+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/nodes/intrinsics/multithreading/LLVMPThreadStart.java
@@ -94,7 +94,7 @@ public final class LLVMPThreadStart {
                         }
                     }
                 }
-                pThreadContext.clearThreadId();
+                pThreadContext.clearThreadID();
             }
         }
     }